DEV: Move client fixture to common area
